Add version 1 of the FlowTimer
This first version of the FlowTimer is a simple solution hacked together
to run without any dependencies. It could be made nicer with the ability
to add meetings and other features.

For now it features the ability to start a break, that is one fifth of
the work time. It has accumulated streaks of long work sessions, that
will unlock extra break time.