Move timer logic to web worker
The timer has shown to experience delays, when the browser window is
in the background. The solution to this is to move the timer logic into
a web worker.

This is a major rewrite, but it should be doable. This first change
simply moves the timer logic into a web worker. All the remaining logic
needs to follow later.